+++ SubTTS/STDVDPlayer.m
@@ -1,9 +1,9 @@
 //
 //  STDVDPlayer.m
-//  SubTTS
+//  Part of SubTTS: Subtitles-to-speech
 //
-//  Created by peter ljunglöf on 2012-03-04.
-//  Copyright (c) 2012 göteborgs universtiet. All rights reserved.
+//  Created by Peter Ljunglöf on 2012-03-04.
+//  Copyright (C) 2012 Peter Ljunglöf. All rights reserved.
 //
 
 #import "STDVDPlayer.h"
@@ -23,7 +23,7 @@
 
 #define HASMOVIE ([[self getPlayer] hasMedia])
 
-- (int) numberOfMovies {
+- (NSInteger) numberOfMovies {
     if (HASMOVIE)
         return 1;
     else
@@ -44,7 +44,7 @@
     }
 }
 
-- (DVDPlayerApplication*) movieAtIndex: (int)nr { 
+- (DVDPlayerApplication*) movieAtIndex: (NSInteger)nr {
     if (nr == 0 && HASMOVIE)
         return [self getPlayer];
     else
