+++ SubTTS/STPlayer.m
@@ -1,9 +1,9 @@
 //
 //  STMoviePlayer.m
-//  SubTTS
+//  Part of SubTTS: Subtitles-to-speech
 //
-//  Created by peter ljunglöf on 2012-02-26.
-//  Copyright (c) 2012 göteborgs universtiet. All rights reserved.
+//  Created by Peter Ljunglöf on 2012-02-26.
+//  Copyright (C) 2012 Peter Ljunglöf. All rights reserved.
 //
 
 #import "STPlayer.h"
@@ -13,9 +13,9 @@
 - (NSString*) title { return nil; }
 - (NSString*) bundleID { return nil; }
 
-- (int) numberOfMovies { return 0; }
+- (NSInteger) numberOfMovies { return 0; }
 - (NSArray*) movies { return nil; }
-- (id) movieAtIndex: (int)nr { return nil; }
+- (id) movieAtIndex: (NSInteger)nr { return nil; }
 - (id) movieWithName: (NSString*)name { return nil; }
 
 - (NSString*) nameOfMovie: (id)movie { return nil; }
@@ -61,7 +61,7 @@
 
 // Some players only give currentTime as integers
 // We guess the fraction by extrapolation from the previous invocation
-- (NSTimeInterval) extrapolateCurrentTime: (int)currentSeconds 
+- (NSTimeInterval) extrapolateCurrentTime: (NSInteger)currentSeconds
                                 isPlaying: (BOOL)isPlaying 
 {
     static NSTimeInterval lastTime = -1;
