+++ SubTTS/STSubtitleArray.h
@@ -1,9 +1,9 @@
 //
 //  STSubtitleArray.h
-//  SubTTS
+//  Part of SubTTS: Subtitles-to-speech
 //
-//  Created by peter ljunglöf on 2012-02-28.
-//  Copyright (c) 2012 göteborgs universtiet. All rights reserved.
+//  Created by Peter Ljunglöf on 2012-02-28.
+//  Copyright (C) 2012 Peter Ljunglöf. All rights reserved.
 //
 
 #import <Foundation/Foundation.h>
@@ -15,9 +15,9 @@
     STSubtitle* finalSubtitle;
 }
 
-- (int) count;
-- (STSubtitle*) subtitle: (int)nr;
-- (int) nextSubtitle: (NSTimeInterval)now;
+- (NSInteger) count;
+- (STSubtitle*) subtitle: (NSInteger)nr;
+- (NSInteger) nextSubtitle: (NSTimeInterval)now;
 
 - (void) addSubtitle: (NSString*)text 
                start: (NSTimeInterval)start
